refactor(desktopui): move events out of stream state
this file is huge ahhhh pulled events out as a start,
but could prob break it up more
